backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
slowapi>=0.1.9
python-docx>=1.1.0
requests>=2.31.0
pydantic>=2.0.0
python-dotenv>=1.0.0
razorpay>=1.4.1
supabase>=2.3.0
